chore(.yarnrc.yml): add npm registries configuration for GitHub Packages and npmjs.org to enable authentication with NODE_AUTH_TOKEN chore(package.json): add repository field to all packages to specify the GitHub repository URL
chore(package.json): add repository field with GitHub repository URL for all packages

The repository field was added to the package.json files of all packages in order to provide a link to the GitHub repository where the packages are hosted. This will make it easier for users and contributors to access the source code and contribute to the project.